#27 - moves components to its module
