+++ core/ui/src/androidTest/java/br/com/rbrthmn/ui/components/ReservesDropdownMenuUITest.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.components
+package br.com.rbrthmn.ui.components
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertIsDisplayed
@@ -26,11 +26,11 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import br.com.rbrthmn.R
-import br.com.rbrthmn.ui.components.ReservesDropdownMenu
+import br.com.rbrthmn.ui.BaseUITest
+import br.com.rbrthmn.ui.R
 import org.junit.Test
 
-class ReservesDropdownMenuTest : br.com.rbrthmn.ui.BaseUITest() {
+class ReservesDropdownMenuTest : BaseUITest() {
     override val composeTestRule = createAndroidComposeRule<ComponentActivity>()
 
     private var selectedReserve: String? = null
