+++ feature/operations/src/androidTest/java/br/com/rbrthmn/operations/OperationsUiTest.kt
@@ -1,24 +1,4 @@
-/*
- *
- * Copyright (C) 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Modifications made by Roberto Kenzo Hamano, 2024
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package br.com.rbrthmn.ui.financialcompanion.screens
+package br.com.rbrthmn.operations
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertIsDisplayed
@@ -32,7 +12,8 @@ import br.com.rbrthmn.ui.BaseUITest
 import org.junit.Test
 
 class OperationsUiTest : BaseUITest() {
-    override val composeTestRule: ComposeContentTestRule = createAndroidComposeRule<ComponentActivity>()
+    override val composeTestRule: ComposeContentTestRule =
+        createAndroidComposeRule<ComponentActivity>()
 
     override fun setup() = composeTestRule.setContent { OperationsScreen() }
